+++ History.md
@@ -3,6 +3,7 @@
 ### Version 4.12.0 [not released yet]
 
 - Bugfix: The diff and line splitting did not work properly with grapheme consisting of multiple code points.
+- Bugfix: Changing the amendment text as admin does not clear the motion's view cache in all cases.
 
 ### Version 4.11.1 (2022-12-10)